pyproject.toml
[project]
name = "nexvec"
version = "1.3.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "google-genai>=1.68.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.27",
    "langchain-google-genai>=1.0",
    "nltk>=3.8",
    "numpy>=1.26",
    "pdfplumber>=0.11",
    "pydantic>=2.6",
    "pytest>=8.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0",
    "uvicorn[standard]>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
